tg_bot/requirements.txt
python-telegram-bot==22.5
psycopg2-binary==2.9.11
python-dotenv==1.2.1
# python-dateutil==2.8.2
# pytz==2023.3